project/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-crontab==0.7.1
django-kombu==0.9.4
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.6.3
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
prompt-toolkit==3.0.19
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
